@@ -10,7 +10,7 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -18,7 +18,10 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  
+  # The version in git does not use null safety now, so until changes 
+  # are merged, building will fail, unless local path is used.
+  cupertino_icons: ^1.0.1
   flutter_icons:
 #    path: ../
      git:
