@@ -9,25 +9,14 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Demo',
-      theme: ThemeData(
-        // This is the theme of your application.
-        //
-        // Try running your application with "flutter run". You'll see the
-        // application has a blue toolbar. Then, without quitting the app, try
-        // changing the primarySwatch below to Colors.green and then invoke
-        // "hot reload" (press "r" in the console where you ran "flutter run",
-        // or simply save your changes to "hot reload" in a Flutter IDE).
-        // Notice that the counter didn't reset back to zero; the application
-        // is not restarted.
-        primarySwatch: Colors.blue,
-      ),
+      theme: ThemeData(primarySwatch: Colors.blue),
       home: MyHomePage(title: 'Flutter Icons'),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
 
   final String title;
 
@@ -36,38 +25,27 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-
   @override
   Widget build(BuildContext context) {
-    // This method is rerun every time setState is called, for instance as done
-    // by the _incrementCounter method above.
-    //
-    // The Flutter framework has been optimized to make rerunning build methods
-    // fast, so that you can just rebuild anything that needs updating rather
-    // than having to individually change instances of widgets.
     return Scaffold(
-      appBar: AppBar(
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
-        title: Text(widget.title),
-      ),
+      appBar: AppBar(title: Text(widget.title)),
       body: Padding(
         padding: const EdgeInsets.all(20.0),
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
             Icon(AntDesign.stepforward),
-//            Icon(Ionicons.ios_search),
-//            Icon(FontAwesome.glass),
-//            Icon(MaterialIcons.ac_unit),
-//            Icon(FontAwesome5.address_book),
-//            Icon(FontAwesome5Solid.address_book),
-//            Icon(FontAwesome5Brands.$500px),
-//            Icon(FlutterIcons.$500px_ent),
-//            Icon(FlutterIcons.$500px_faw5d),
-//            Icon(WeatherIcons.wi_alien),
-//            Icon(WeatherIcons.wi_cloud),
-//            Icon(WeatherIcons.wi_day_rain)
+            Icon(Ionicons.ios_search),
+            Icon(FontAwesome.glass),
+            Icon(MaterialIcons.ac_unit),
+            Icon(FontAwesome5.address_book),
+            Icon(FontAwesome5Solid.address_book),
+            Icon(FontAwesome5Brands.$500px),
+            Icon(FlutterIcons.$500px_ent),
+            Icon(FlutterIcons.$500px_faw5d),
+            Icon(WeatherIcons.wi_alien),
+            Icon(WeatherIcons.wi_cloud),
+            Icon(WeatherIcons.wi_day_rain)
           ],
         ),
       ),
